Prepared for State Machine
I think I have all the parts in here, but you probably want to double
check. The pins need to actually be set to numbers, but everything looks
fine. Might need to change the states a bit depends on how we tackle the
second reactor but I think we could just have variable that gets flagged
and changes which reactor.